Add features for desktop-icons, gnome-software, snap-store and ubuntu-dock, check dependencies before installing extra packages in feature scripts, check if flatpak/snap are installed before installing flatpak/snap features, reduce priority of gnome-console as x-terminal-emulator alternative to match gnome-terminal, remove repositories/packages/features before enabling/installing/adding new ones

+++ packages/bleux-features/features/flatpak/thunderbird.sh
@@ -14,6 +14,10 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
+if ! dpkg-query -f '${db:Status-abbrev}' -W flatpak 2> /dev/null | grep -q '^.i'; then
+    exit 1
+fi
+
 case $1 in
     add)
         flatpak install -y --noninteractive --no-related org.mozilla.Thunderbird
